requirements.txt
configparser>=5.0.2
python-telegram-bot>=13.6